requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
future==0.18.3
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==22.0
pefile==2022.5.30
Pillow==9.4.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PySocks==1.7.1
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tk==0.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
ttkwidgets==0.13.0
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
XlsxWriter==3.0.6
